+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - ...
 
-## [1.8.0] - 2025-07-xx
+## [1.8.0] - 2025-07-30
 
 ### Added
 - Supporting the 'ln', 'exp', 'sqrt' and 'power' math functions.